add search function for reagentkits and reagentlots separate descriptors from entities add ability to create entries in basic descriptors add create function in entities to create new Entities from scratch and post them to the LIMS.
