+++ tests/test_entities.py
@@ -4,9 +4,8 @@ from sys import version_info
 from io import BytesIO
 
 from genologics.lims import Lims
-from genologics.entities import StringDescriptor, StringAttributeDescriptor, StringListDescriptor, \
-    StringDictionaryDescriptor, IntegerDescriptor, BooleanDescriptor, UdfDictionary, StepActions, Researcher, Artifact, \
-    Step, StepPlacements, Container
+from genologics.entities import StepActions, Researcher, Artifact, \
+    Step, StepPlacements, Container, Stage, ReagentKit, ReagentLot
 
 if version_info.major == 2:
     from mock import patch, Mock
@@ -14,241 +13,76 @@ else:
     from unittest.mock import patch, Mock
 
 
+url = 'http://testgenologics.com:4040'
 
+########
+# Entities in XML
+generic_artifact_xml = """<?xml version='1.0' encoding='utf-8'?>
+<art:artifact xmlns:art="http://genologics.com/ri/artifact"  xmlns:file="http://genologics.com/ri/file" xmlns:udf="http://genologics.com/ri/userdefined"  uri="{url}/api/v2/artifacts/a1" limsid="a1">
+<name>test_sample1</name>
+<type>Analyte</type>
+<output-type>Analyte</output-type>
+<qc-flag>PASSED</qc-flag>
+<location>
+  <container uri="{url}/api/v2/containers/c1" limsid="c1"/>
+  <value>A:1</value>
+</location>
+<working-flag>true</working-flag>
+<sample uri="{url}/api/v2/samples/s1" limsid="s1"/>
+<udf:field type="Numeric" name="Ave. Conc. (ng/uL)">1</udf:field>
+<udf:field type="String" name="Workflow Desired">TruSeq Nano DNA Sample Prep</udf:field>
+<workflow-stages>
+<workflow-stage status="QUEUED" name="Test workflow s2" uri="{url}/api/v2/configuration/workflows/1/stages/2"/>
+<workflow-stage status="COMPLETE" name="Test workflow s1" uri="{url}/api/v2/configuration/workflows/1/stages/1"/>
+</workflow-stages>
+</art:artifact>"""
 
+generic_step_placements_xml = """<?xml version='1.0' encoding='utf-8'?>
+<stp:placements xmlns:stp="http://genologics.com/ri/step" uri="{url}/steps/s1/placements">
+  <step uri="{url}/steps/s1" />
+  <configuration uri="{url}/configuration/protocols/1/steps/1">Step name</configuration>
+  <selected-containers>
+    <container uri="{url}/containers/{container}" />
+  </selected-containers>
+  <output-placements>
+    <output-placement uri="{url}/artifacts/a1">
+      <location>
+        <container limsid="{container}" uri="{url}/containers/{container}" />
+        <value>{loc1}</value>
+      </location>
+    </output-placement>
+    <output-placement uri="{url}/artifacts/a2">
+      <location>
+        <container limsid="{container}" uri="{url}/containers/{container}" />
+        <value>{loc2}</value>
+      </location>
+    </output-placement>
+  </output-placements>
+</stp:placements>"""
 
-class TestEntities(TestCase):
+generic_reagentkit_xml = """<?xml version='1.0' encoding='utf-8'?>
+<kit:reagent-kit xmlns:kit="http://genologics.com/ri/reagentkit" uri="{url}:8080/api/v2/reagentkits/r1">
+<name>regaentkitname</name>
+<supplier>reagentProvider</supplier>
+<website>www.reagentprovider.com</website>
+<archived>false</archived>
+</kit:reagent-kit>"""
 
-    def test_pass(self):
-        pass
+generic_reagentlot_xml = """<?xml version='1.0' encoding='utf-8'?>
+<lot:reagent-lot xmlns:lot="http://genologics.com/ri/reagentlot" limsid="l1" uri="{url}/api/v2/reagentlots/l1">
+<reagent-kit uri="{url}/api/v2/reagentkits/r1" name="kitname"/>
+<name>kitname</name>
+<lot-number>100</lot-number>
+<created-date>2015-07-16</created-date>
+<last-modified-date>2015-08-17</last-modified-date>
+<expiry-date>2022-08-16</expiry-date>
+<created-by uri="{url}/api/v2/researchers/1"/>
+<last-modified-by uri="{url}/api/v2/researchers/1"/>
+<status>ARCHIVED</status>
+<usage-count>1</usage-count>
+</lot:reagent-lot>"""
 
-
-class TestDescriptor(TestCase):
-
-    def _make_desc(self, klass, *args, **kwargs):
-        return klass(*args, **kwargs)
-
-    def _tostring(self, e):
-        outfile = BytesIO()
-        ElementTree.ElementTree(e).write(outfile, encoding='utf-8', xml_declaration=True)
-        return outfile.getvalue()
-
-
-class TestStringDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<name>test name</name>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringDescriptor, 'name')
-        assert sd.__get__(self.instance, None) == "test name"
-
-    def test__set__(self):
-        sd = self._make_desc(StringDescriptor, 'name')
-        sd.__set__(self.instance, "new test name")
-        assert self.et.find('name').text == "new test name"
-
-
-class TestStringAttributeDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry name="test name">
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringAttributeDescriptor, 'name')
-        assert sd.__get__(self.instance, None) == "test name"
-
-
-class TestStringListDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<test-subentry>A01</test-subentry>
-<test-subentry>B01</test-subentry>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringListDescriptor, 'test-subentry')
-        assert sd.__get__(self.instance, None) == ['A01', 'B01']
-
-
-class TestStringDictionaryDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<test-subentry>
-<test-firstkey/>
-<test-secondkey>second value</test-secondkey>
-</test-subentry>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringDictionaryDescriptor, 'test-subentry')
-        res = sd.__get__(self.instance, None)
-        assert type(res) == dict
-        self.assertIsNone(res['test-firstkey'])
-        assert res['test-secondkey'] == 'second value'
-
-class TestIntegerDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<count>32</count>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(IntegerDescriptor, 'count')
-        assert sd.__get__(self.instance, None) == 32
-
-    def test__set__(self):
-        sd = self._make_desc(IntegerDescriptor, 'count')
-        sd.__set__(self.instance, 23)
-        assert self.et.find('count').text == 23
-        # FIXME: The BooleanDescriptor (and the IntegerDescriptor) uses the StringDescriptor
-        # Using them with their expected type makes serialization crash
-
-class TestBooleanDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<istest>true</istest>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(BooleanDescriptor, 'istest')
-        assert sd.__get__(self.instance, None) == True
-
-    def test__set__(self):
-        # FIXME: The BooleanDescriptor (and the IntegerDescriptor) uses the StringDescriptor
-        # Using them with their expected type makes serialization crash
-        sd = self._make_desc(BooleanDescriptor, 'istest')
-        sd.__set__(self.instance, False)
-        assert self.et.find('istest').text == False
-        #sd.__set__(self.instance, True)
-        #print(self._tostring(self.et))
-        #sd.__get__(self.instance, None)
-
-
-class TestUdfDictionary(TestCase):
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry xmlns:udf="http://genologics.com/ri/userdefined">
-<udf:field type="String" name="test">stuff</udf:field>
-<udf:field type="Numeric" name="how much">42</udf:field>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-        self.dict1 = UdfDictionary(self.instance)
-        self.dict2 = UdfDictionary(self.instance, udt=True)
-
-    def _get_udf_value(self, udf_dict, key):
-        for e in udf_dict._elems:
-            if e.attrib['name'] != key:
-                continue
-            else:
-                return e.text
-
-    def test_get_udt(self):
-        pass
-
-    def test_set_udt(self):
-        pass
-
-    def test__update_elems(self):
-        pass
-
-    def test__prepare_lookup(self):
-        pass
-
-    def test___contains__(self):
-        pass
-
-    def test___getitem__(self):
-        pass
-
-    def test___setitem__(self):
-        assert self._get_udf_value(self.dict1, 'test') == 'stuff'
-        self.dict1.__setitem__('test', 'other')
-        assert self._get_udf_value(self.dict1, 'test') == 'other'
-
-        assert self._get_udf_value(self.dict1, 'how much') == '42'
-        self.dict1.__setitem__('how much', 21)
-        assert self._get_udf_value(self.dict1, 'how much') == '21'
-
-        self.assertRaises(TypeError, self.dict1.__setitem__, 'how much', '433')
-
-        #I'm not sure if this is the expected behaviour
-        self.dict1.__setitem__('how much', None)
-        assert self._get_udf_value(self.dict1, 'how much') == b'None'
-
-    def test___setitem__unicode(self):
-        assert self._get_udf_value(self.dict1, 'test') == 'stuff'
-        self.dict1.__setitem__('test', u'unicode')
-        assert self._get_udf_value(self.dict1, 'test') == 'unicode'
-
-        self.dict1.__setitem__(u'test', 'unicode2')
-        assert self._get_udf_value(self.dict1, 'test') == 'unicode2'
-
-
-
-    def test___delitem__(self):
-        pass
-
-    def test_items(self):
-        pass
-
-    def test_clear(self):
-        pass
-
-    def test___iter__(self):
-        pass
-
-    def test___next__(self):
-        pass
-
-    def test_get(self):
-        pass
-
-def elements_equal(e1, e2):
-    if e1.tag != e2.tag: return False
-    if e1.text and e2.text and e1.text.strip() != e2.text.strip(): return False
-    if e1.tail and e2.tail and e1.tail.strip() != e2.tail.strip(): return False
-    if e1.attrib != e2.attrib: return False
-    if len(e1) != len(e2): return False
-    return all(elements_equal(c1, c2) for c1, c2 in zip(e1, e2))
-
-class TestEntities(TestCase):
-    url = 'http://testgenologics.com:4040'
-    dummy_xml="""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-    <dummy></dummy>"""
-
-    def setUp(self):
-        self.lims = Lims(self.url, username='test', password='password')
-
-    def _tostring(self, entity):
-        return self.lims.tostring(ElementTree.ElementTree(entity.root)).decode("utf-8")
-
-
-class TestStepActions(TestEntities):
-    url = 'http://testgenologics.com:4040'
-    step_actions_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/step" uri="...">
+generic_step_actions_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/step" uri="...">
   <step rel="..." uri="{url}/steps/s1">
   </step>
   <configuration uri="{url}/config/1">...</configuration>
@@ -282,9 +116,10 @@ class TestStepActions(TestEntities):
       </escalated-artifact>
     </escalated-artifacts>
   </escalation>
-</stp:actions>""".format(url=url)
+</stp:actions>"""
 
-    step_actions_no_escalation_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/step" uri="...">
+
+generic_step_actions_no_escalation_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/step" uri="...">
   <step rel="..." uri="{url}/steps/s1">
   </step>
   <configuration uri="{url}/config/1">...</configuration>
@@ -292,7 +127,36 @@ class TestStepActions(TestEntities):
     <next-action artifact-uri="{url}/artifacts/a1" action="requeue" step-uri="{url}/steps/s1" rework-step-uri="{url}/steps/s2">
     </next-action>
   </next-actions>
-</stp:actions>""".format(url=url)
+</stp:actions>"""
+
+class TestEntities(TestCase):
+
+    def test_pass(self):
+        pass
+
+
+def elements_equal(e1, e2):
+    if e1.tag != e2.tag: return False
+    if e1.text and e2.text and e1.text.strip() != e2.text.strip(): return False
+    if e1.tail and e2.tail and e1.tail.strip() != e2.tail.strip(): return False
+    if e1.attrib != e2.attrib: return False
+    if len(e1) != len(e2): return False
+    return all(elements_equal(c1, c2) for c1, c2 in zip(e1, e2))
+
+class TestEntities(TestCase):
+    dummy_xml="""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+    <dummy></dummy>"""
+
+    def setUp(self):
+        self.lims = Lims(url, username='test', password='password')
+
+    def _tostring(self, entity):
+        return self.lims.tostring(ElementTree.ElementTree(entity.root)).decode("utf-8")
+
+
+class TestStepActions(TestEntities):
+    step_actions_xml = generic_step_actions_xml.format(url=url)
+    step_actions_no_escalation_xml = generic_step_actions_no_escalation_xml.format(url=url)
 
     def test_escalation(self):
         s = StepActions(uri=self.lims.get_uri('steps', 'step_id', 'actions'), lims=self.lims)
@@ -321,33 +185,10 @@ class TestStepActions(TestEntities):
 
 
 class TestStepPlacements(TestEntities):
-    url = 'http://testgenologics.com:4040'
-    generic_step_placements_xml = """<?xml version='1.0' encoding='utf-8'?>
-<stp:placements xmlns:stp="http://genologics.com/ri/step" uri="{url}/steps/s1/placements">
-  <step uri="{url}/steps/s1" />
-  <configuration uri="{url}/configuration/protocols/1/steps/1">Step name</configuration>
-  <selected-containers>
-    <container uri="{url}/containers/{container}" />
-  </selected-containers>
-  <output-placements>
-    <output-placement uri="{url}/artifacts/a1">
-      <location>
-        <container limsid="{container}" uri="{url}/containers/{container}" />
-        <value>{loc1}</value>
-      </location>
-    </output-placement>
-    <output-placement uri="{url}/artifacts/a2">
-      <location>
-        <container limsid="{container}" uri="{url}/containers/{container}" />
-        <value>{loc2}</value>
-      </location>
-    </output-placement>
-  </output-placements>
-</stp:placements>"""
+
     original_step_placements_xml = generic_step_placements_xml.format(url=url, container="c1", loc1='1:1', loc2='2:1')
     modloc_step_placements_xml = generic_step_placements_xml.format(url=url, container="c1", loc1='3:1', loc2='4:1')
     modcont_step_placements_xml = generic_step_placements_xml.format(url=url, container="c2", loc1='1:1', loc2='1:1')
-
 
 
     def test_get_placements_list(self):
@@ -381,3 +222,70 @@ class TestStepPlacements(TestEntities):
             new_placements = [[a1,(c2,'1:1')], [a2,(c2,'1:1')]]
             s.set_placement_list(new_placements)
             assert elements_equal(s.root, ElementTree.fromstring(self.modcont_step_placements_xml))
+
+
+
+
+class TestArtifacts(TestEntities):
+
+    root_artifact_xml = generic_artifact_xml.format(url=url)
+
+    def test_input_artifact_list(self):
+        a = Artifact(uri=self.lims.get_uri('artifacts', 'a1'), lims=self.lims)
+        with patch('requests.Session.get',return_value=Mock(content = self.root_artifact_xml, status_code=200)):
+            assert a.input_artifact_list() == []
+
+    def test_workflow_stages_and_statuses(self):
+        a = Artifact(uri=self.lims.get_uri('artifacts', 'a1'), lims=self.lims)
+        expected_wf_stage = [
+            (Stage(self.lims, uri = url + '/api/v2/configuration/workflows/1/stages/2'), 'QUEUED', 'Test workflow s2'),
+            (Stage(self.lims, uri = url + '/api/v2/configuration/workflows/1/stages/1'), 'COMPLETE', 'Test workflow s1')
+        ]
+        with patch('requests.Session.get',return_value=Mock(content = self.root_artifact_xml, status_code=200)):
+            assert a.workflow_stages_and_statuses == expected_wf_stage
+
+
+
+
+class TestReagentKits(TestEntities):
+    url = 'http://testgenologics.com:4040'
+    reagentkit_xml = generic_reagentkit_xml.format(url=url)
+
+    def test_parse_entity(self):
+        r = ReagentKit(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
+        with patch('requests.Session.get', return_value=Mock(content = self.reagentkit_xml, status_code=200)):
+            assert r.name == 'regaentkitname'
+            assert r.supplier == 'reagentProvider'
+            assert r.website == 'www.reagentprovider.com'
+            assert r.archived == False
+
+    def test_create_entity(self):
+        with patch('genologics.lims.requests.post', return_value=Mock(content = self.reagentkit_xml, status_code=201)):
+            r = ReagentKit.create(self.lims, name='regaentkitname', supplier='reagentProvider', website='www.reagentprovider.com', archived=False)
+        self.assertRaises(TypeError,ReagentKit.create,self.lims, error='test')
+
+
+class TestReagentLots(TestEntities):
+    reagentlot_xml = generic_reagentlot_xml.format(url=url)
+    reagentkit_xml = generic_reagentkit_xml.format(url=url)
+
+    def test_parse_entity(self):
+        l = ReagentLot(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
+        with patch('requests.Session.get', return_value=Mock(content = self.reagentlot_xml, status_code=200)):
+            assert l.name == 'kitname'
+            assert l.lot_number == '100'
+            assert l.status == 'ARCHIVED'
+
+    def test_create_entity(self):
+         with patch('requests.Session.get', return_value=Mock(content = self.reagentkit_xml, status_code=200)):
+             r = ReagentKit(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
+         with patch('genologics.lims.requests.post', return_value=Mock(content = self.reagentlot_xml, status_code=201)) as patch_post:
+             l = ReagentLot.create(
+                    self.lims,
+                    reagent_kit=r,
+                    name='kitname',
+                    lot_number='99',
+                    expiry_date='2020-05-01',
+                    status='ACTIVE'
+             )
+             print(patch_post.call_args_list)
